# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VirJoy AI — prompt-based AI video creation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Detected or specified video style
 */
export type VideoJobVideoType = typeof VideoJobVideoType[keyof typeof VideoJobVideoType];


export const VideoJobVideoType = {
  ad: 'ad',
  horror: 'horror',
  promo: 'promo',
  vlog: 'vlog',
} as const;

export type VideoJobPlan = typeof VideoJobPlan[keyof typeof VideoJobPlan];


export const VideoJobPlan = {
  free: 'free',
  starter: 'starter',
  creator: 'creator',
  premium: 'premium',
} as const;

export type VideoJobStatus = typeof VideoJobStatus[keyof typeof VideoJobStatus];


export const VideoJobStatus = {
  queued: 'queued',
  processing: 'processing',
  done: 'done',
  failed: 'failed',
} as const;

export interface VideoJob {
  id: string;
  /**
     * The user's text prompt describing the video
     * @nullable
     */
  prompt?: string | null;
  /** @nullable */
  title?: string | null;
  /** Detected or specified video style */
  videoType: VideoJobVideoType;
  /** Duration in seconds (10, 30, 60, 180) */
  duration: number;
  plan: VideoJobPlan;
  status: VideoJobStatus;
  hasWatermark: boolean;
  /** @nullable */
  outputUrl?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  imageCount?: number;
  clipCount?: number;
  createdAt: string;
  updatedAt?: string;
}

/**
 * Optional — auto-detected from prompt if omitted
 */
export type VideoJobInputVideoType = typeof VideoJobInputVideoType[keyof typeof VideoJobInputVideoType];


export const VideoJobInputVideoType = {
  ad: 'ad',
  horror: 'horror',
  promo: 'promo',
  vlog: 'vlog',
} as const;

export type VideoJobInputPlan = typeof VideoJobInputPlan[keyof typeof VideoJobInputPlan];


export const VideoJobInputPlan = {
  free: 'free',
  starter: 'starter',
  creator: 'creator',
  premium: 'premium',
} as const;

export interface VideoJobInput {
  /** Text prompt describing the video. The system auto-detects style from this. */
  prompt?: string;
  /** Optional — auto-detected from prompt if omitted */
  videoType?: VideoJobInputVideoType;
  /** Duration in seconds */
  duration: number;
  plan: VideoJobInputPlan;
  images?: Blob[];
  clips?: Blob[];
}

export type VideosSummaryByStatus = {[key: string]: number};

export type VideosSummaryByType = {[key: string]: number};

export type VideosSummaryByPlan = {[key: string]: number};

export interface VideosSummary {
  total: number;
  byStatus: VideosSummaryByStatus;
  byType: VideosSummaryByType;
  byPlan: VideosSummaryByPlan;
}

export type PlanId = typeof PlanId[keyof typeof PlanId];


export const PlanId = {
  free: 'free',
  starter: 'starter',
  creator: 'creator',
  premium: 'premium',
} as const;

export type PlanQuality = typeof PlanQuality[keyof typeof PlanQuality];


export const PlanQuality = {
  low: 'low',
  standard: 'standard',
  high: 'high',
} as const;

export interface Plan {
  id: PlanId;
  name: string;
  /** Price in INR (0 = free) */
  price: number;
  maxImages: number;
  maxClips: number;
  /** Max duration in seconds */
  maxDuration: number;
  watermark: boolean;
  quality: PlanQuality;
  aiStory: boolean;
  enhancedCinematicEffects?: boolean;
  priorityRendering?: boolean;
  ideaToVideo?: boolean;
  description?: string;
}

export interface StoryInput {
  /** The user's prompt describing the video idea */
  prompt: string;
}

export type StoryResultDetectedType = typeof StoryResultDetectedType[keyof typeof StoryResultDetectedType];


export const StoryResultDetectedType = {
  ad: 'ad',
  horror: 'horror',
  promo: 'promo',
  vlog: 'vlog',
} as const;

export interface Scene {
  index: number;
  description: string;
  duration: number;
  effect: string;
}

export interface StoryResult {
  title: string;
  story: string;
  detectedType: StoryResultDetectedType;
  scenes: Scene[];
}

export interface DeleteResult {
  success: boolean;
}

export interface ApiError {
  error: string;
}