# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VirJoy AI — prompt-based AI video creation API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanId } from './planId';
import type { PlanQuality } from './planQuality';

export interface Plan {
  id: PlanId;
  name: string;
  /** Price in INR (0 = free) */
  price: number;
  maxImages: number;
  maxClips: number;
  /** Max duration in seconds */
  maxDuration: number;
  watermark: boolean;
  quality: PlanQuality;
  aiStory: boolean;
  enhancedCinematicEffects?: boolean;
  priorityRendering?: boolean;
  ideaToVideo?: boolean;
  description?: string;
}